refactor : 도메인 패키지 구조 수정
