+++ meerket/meerket-domain/src/main/java/org/j1p5/domain/auth/OauthSender.java
@@ -1,5 +1,7 @@
 package org.j1p5.domain.auth;
 
+import org.j1p5.domain.auth.dto.OauthProfile;
+import org.j1p5.domain.auth.dto.OauthToken;
 import org.j1p5.domain.user.entity.Provider;
 import org.springframework.stereotype.Component;
