+++ meerket/meerket-domain/src/main/java/org/j1p5/domain/auth/OauthService.java
@@ -1,6 +1,8 @@
 package org.j1p5.domain.auth;
 
 import lombok.RequiredArgsConstructor;
+import org.j1p5.domain.auth.dto.OauthProfile;
+import org.j1p5.domain.auth.validator.LoginValidator;
 import org.j1p5.domain.user.entity.Provider;
 import org.j1p5.domain.user.entity.UserEntity;
 import org.j1p5.domain.user.service.UserAppender;
