+++ meerket/meerket-domain/src/main/java/org/j1p5/domain/auth/dto/NaverProfile.java
@@ -1,4 +1,4 @@
-package org.j1p5.domain.auth;
+package org.j1p5.domain.auth.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
