+++ meerket/meerket-domain/src/main/java/org/j1p5/domain/auth/dto/OauthProfile.java
@@ -1,4 +1,4 @@
-package org.j1p5.domain.auth;
+package org.j1p5.domain.auth.dto;
 
 public interface OauthProfile {
     String getId();
