+++ meerket/meerket-infrastructure/src/main/java/org/j1p5/infrastructure/kakao/KakaoAuthClient.java
@@ -1,6 +1,6 @@
 package org.j1p5.infrastructure.kakao;
 
-import org.j1p5.domain.auth.KakaoToken;
+import org.j1p5.domain.auth.dto.KakaoToken;
 import org.j1p5.infrastructure.global.config.FeignClientConfig;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
