+++ meerket/meerket-infrastructure/src/main/java/org/j1p5/infrastructure/kakao/KakaoProfileClient.java
@@ -1,6 +1,6 @@
 package org.j1p5.infrastructure.kakao;
 
-import org.j1p5.domain.auth.KakaoProfile;
+import org.j1p5.domain.auth.dto.KakaoProfile;
 import org.j1p5.infrastructure.global.config.FeignClientConfig;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
