+++ meerket/meerket-infrastructure/src/main/java/org/j1p5/infrastructure/naver/NaverProfileClient.java
@@ -1,6 +1,6 @@
 package org.j1p5.infrastructure.naver;
 
-import org.j1p5.domain.auth.NaverProfile;
+import org.j1p5.domain.auth.dto.NaverProfile;
 import org.j1p5.infrastructure.global.config.FeignClientConfig;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
